Bugfix + moved random game choosing in GameEngine.
Moved logic for getting a random word and the array of words out of
Hangman and in GameEngine. Hangaman's constructor now takes one argument
- the word. Private property with validity checks for the word
introduced. Moved the logic of Initialize method in the constructor and
removed the now obsolete method.
Bugfix in SocreBoard's GetWorstTopScore() method where the logic tried
to asses highScores[highScores.Count - 1] but there were no record in
highScores (IndexOutOfRange for index=-1).